Fix test for composable recorder with parameter override option
