+++ rosbag2_transport/CMakeLists.txt
@@ -234,7 +234,7 @@ function(create_tests_for_rmw_implementation)
     AMENT_DEPS test_msgs rosbag2_test_common)
 
   rosbag2_transport_add_gmock(test_component_parameters
-    test/rosbag2_transport/test_component_parameters.cpp
+      test/rosbag2_transport/test_composable_recorder.cpp
     LINK_LIBS rosbag2_transport
     AMENT_DEPS test_msgs rosbag2_test_common)
